Update BehaviorDb.TheEther.cs
Ok. i believe i did this correct. so just read over the places i changed just to make sure >.<

I changed the Ether skeleton shots from 5 to 3.         5 can easily stack up and kill

I changed the delay on the Vortex shots for lord of death.

The code is not here so you are going to have to do this for me.   *look below*

Change the boomerang shots damage to 80.

Change the Vortex shots damage to 100